+++ src/backend/joanie/core/models/courses.py
@@ -926,6 +926,102 @@ class CourseProductRelation(BaseModel):
             "has_seats_left": has_seats_left or not offering_rule_is_blocking,
         }
 
+    def activate_deep_links(self, is_active=True):
+        """
+        Activate or deactivate the deeplinks of the offering.
+        """
+        self.organization_links.update(is_active=is_active)
+
+    @property
+    def has_deep_links(self) -> bool:
+        """Return boolean value whether the offering has deep links declared"""
+        return self.organization_links.exists()
+
+    @property
+    def can_access_deep_links(self) -> bool:
+        """Return boolean whether the deep links of the offering are activated"""
+        return self.organization_links.filter(is_active=True).exists()
+
+
+class OfferingDeepLink(BaseModel):
+    """
+    OfferingDeepLink represents external links that allows a learner to subscribe to a course
+    from an external platform (e.g : Mon Compte Formation in France). This model gives an
+    organization the possibility to add their deep link for one specific offering.
+    """
+
+    deep_link = models.URLField(
+        _("deep_link"),
+        help_text=_(
+            "Deep link of an offering where a learner can subscribe with an external link."
+        ),
+        unique=True,
+        blank=False,
+        null=False,
+    )
+    offering = models.ForeignKey(
+        to=CourseProductRelation,
+        on_delete=models.CASCADE,
+        verbose_name=_("offering"),
+        related_name="organization_links",
+    )
+    organization = models.ForeignKey(
+        to=Organization,
+        verbose_name=_("organization"),
+        related_name="offering_links",
+        on_delete=models.CASCADE,
+    )
+    is_active = models.BooleanField(
+        verbose_name=_("is active"),
+        default=False,
+        help_text=_("Activate or deactivate deep link of an offering"),
+    )
+
+    class Meta:
+        db_table = "joanie_offering_deep_link"
+        verbose_name = _("Offering Deep Link")
+        verbose_name_plural = _("Offering Deep Links")
+        ordering = ["-created_on"]
+        constraints = [
+            models.UniqueConstraint(
+                fields=["organization", "offering"],
+                name="unique_offering_per_organization",
+            ),
+        ]
+
+    def __str__(self):
+        return f"{self.offering.course}: {self.organization} deep link"
+
+    def clean(self):
+        """
+        Ensure the organization is related to the offering before creating the object, otherwise
+        raise a validation error. Deep links are exclusive to products of type credentials.
+        """
+        if not self.offering.organizations.filter(pk=self.organization.id).exists():
+            raise ValidationError(_("Organization should be related to the offering."))
+
+        if self.offering.product.type != enums.PRODUCT_TYPE_CREDENTIAL:  # pylint:disable=no-member
+            raise ValidationError(
+                _("Only product type credentials are allowed to have deep link")
+            )
+
+        return super().clean()
+
+    def save(self, *args, **kwargs):
+        """Enforce validation each time an instance is saved."""
+        self.full_clean()
+        super().save(*args, **kwargs)
+
+    def delete(self, using=None, keep_parents=False):
+        """
+        A deep link can be deleted if it's not active only.
+        """
+        if self.is_active:
+            raise ValidationError(
+                _("You cannot delete this offering deep link, it's active.")
+            )
+        return super().delete(using=using, keep_parents=keep_parents)
+
 
 class CourseRun(parler_models.TranslatableModel, BaseModel):
     """